project continuation

+++ src/main/java/com/shopping_list/controller/ShoppingController.java
@@ -1,8 +1,10 @@
 package com.shopping_list.controller;
 
 import com.shopping_list.Repository.ShoppingRepository;
+import com.shopping_list.Repository.TaskRepository;
 import com.shopping_list.entities.Shopping;
 import com.shopping_list.entities.Task;
+import com.shopping_list.messages.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -14,13 +16,15 @@ import java.util.Optional;
 
 
 @Controller
-@RequestMapping("/shopping")
 public class ShoppingController {
 
     @Autowired
     private ShoppingRepository shoppingRepository;
 
-    @GetMapping("/all")
+    @Autowired
+    private TaskRepository taskRepository;
+
+    @GetMapping("/shopping/all")
     public String findAll(Model model){
         model.addAttribute("shoppings", shoppingRepository.findAll());
         model.addAttribute("tasks", new Task());
@@ -28,25 +32,25 @@ public class ShoppingController {
         return "shopping/shoppings";
     }
 
-    @GetMapping("/form")
+    @GetMapping("/shopping/form")
     public String form(Model model){
         model.addAttribute("shopping", new Shopping());
         return "shopping/form";
     }
 
-    @PostMapping("/save")
+    @PostMapping("/shopping/save")
     public String save(Shopping shopping){
         shoppingRepository.save(shopping);
         return "shopping/redirection";
     }
 
-    @GetMapping("/update/{id_shop}")
+    @GetMapping("/shopping/update/{id_shop}")
     public String updatedShopping(Model model, @PathVariable Long id_shop){
       Optional<Shopping> optional =  shoppingRepository.findById(id_shop);
         model.addAttribute("shopping", optional.get());
         return "shopping/update";
     }
-    @PostMapping("/update/{id_shop}")
+    @PostMapping("/shopping/update/{id_shop}")
     public String save(Shopping shopping, @PathVariable Long id_shop, BindingResult result){
 
         if(result.hasErrors()) {
@@ -57,16 +61,20 @@ public class ShoppingController {
         return "redirect:/shopping/detail/" +shop.getId_shop() ;
     }
 
-    @GetMapping("/delete/{id_shop}")
-    public String deleteById(@PathVariable Long id_shop, Model model) {
-        Shopping shopping = shoppingRepository.findById(id_shop)
-                .orElseThrow(() -> new IllegalArgumentException("Invalid shopping id:" +id_shop));
-        shoppingRepository.delete(shopping);
-        model.addAttribute("shoppings", shoppingRepository.findAll());
-        return "redirect:/";
+    @DeleteMapping("/task/{task_id}/shoppings/{id_shop}")
+    public String deleteById(@PathVariable Long id_shop, @PathVariable Long task_id, Model model) {
+        if(!taskRepository.existsById(task_id)) {
+            throw new NotFoundException("Task not found with id " + task_id);
+        }
+
+        return shoppingRepository.findById(id_shop)
+                .map(shopping -> {
+                    shoppingRepository.delete(shopping);
+                    return "redirect:/";
+                }).orElseThrow(() -> new NotFoundException("shopping not found with id" + id_shop));
     }
 
-    @GetMapping("/detail/{id_shop}")
+    @GetMapping("/shopping/detail/{id_shop}")
     public String getShop(Model model, @PathVariable Long id_shop){
         Optional<Shopping> optional=shoppingRepository.findById(id_shop);
         model.addAttribute("shopping", optional.get());
@@ -74,14 +82,14 @@ public class ShoppingController {
 
     }
 
-    @GetMapping("/utilisateur/{user_id}")
+    @GetMapping("/shopping/utilisateur/{user_id}")
     public String findByUser(@PathVariable Long user_id, Model model){
         List<Shopping> shoppings = shoppingRepository.findAllByUtilisateurOrderById_shopDesc(user_id);
         model.addAttribute("shoppings", shoppings);
         return "shopping/utilisateur/shoppings";
     }
 
-    @GetMapping("/status/{status_id}")
+    @GetMapping("/shopping/status/{status_id}")
     public String findByStatus(@PathVariable Long status_id, Model model){
         List<Shopping>shoppings = shoppingRepository.findAllByStatusOrderById_shopDesc(status_id);
         model.addAttribute("shoppings", shoppings);
